Getting started with application
A fairly large part of the application was written because most functions had the same idea, which could be easily implemented by the database functions previously written. I should also mention that I implemented most of the functions as static because the main concept of the application part is a bit far from object-oriented programming because we have a large database and practically do not store any objects in the class and to access the information needed, We refer to the database, so it was wise to choose the functions as static functions. I should also mention that small changes were made to the database functions so that it could better communicate with the application section.