Completing crucial parts of application
In general, I continued to complete the functions required for a banking system replica, an important function that I created in this section was a function to apply for a loan. The function works by depositing an amount as a loan to the applicant's account and this amount is automatically deducted from the individual's account during periods. The challenge for this function was that the operation of this function should not interfere with the program, so the use of python threading functions was a good choice for this part of the program.